requirements.txt
-i https://pypi.python.org/simple
asgiref==3.4.1
boto3==1.17.49
botocore==1.20.112
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1 ; python_full_version >= '3.6.1'
charset-normalizer==2.0.10
cryptography==3.4.7
decorator==4.4.2
deprecated==1.2.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
distlib==0.3.6
django==3.2
django-cors-headers==3.7.0
django-crontab==0.7.1
django-environ==0.9.0
django-storages==1.11.1
djangorestframework==3.12.4
filelock==3.12.2 ; python_version >= '3.7'
identify==2.5.24 ; python_version >= '3.7'
idna==3.3
imageio==2.14.0
imageio-ffmpeg==0.4.5
jmespath==0.10.0
moviepy==1.0.3
nodeenv==1.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==1.22.1
pandas==1.5.3
pillow==8.3.2
platformdirs==3.9.1 ; python_version >= '3.7'
pre-commit==2.20.0
proglog==0.1.9
psycopg2==2.9.1
pycparser==2.21
pycurl==7.43.0.5
pyfcm==1.5.3
pygithub==1.59.0
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
pynacl==1.5.0 ; python_version >= '3.6'
python-dateutil==2.8.2
pytz==2021.3
pyyaml==6.0 ; python_version >= '3.6'
requests==2.27.1
s3transfer==0.3.7
setuptools==68.0.0 ; python_version >= '3.7'
six==1.16.0
sqlparse==0.4.2
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqdm==4.62.3
urllib3==1.26.8
virtualenv==20.24.0 ; python_version >= '3.7'
wrapt==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
